requirements.txt
fastexcel==0.13.0
Flask==3.1.0
openpyxl==3.2.0b1
polars-lts-cpu==1.26.0
